pyproject.toml
[tool.poetry]
name = "asbd"
version = "0.1.0"
description = ""
authors = ["Ruslan Sirazhetdinov <ruslansir@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
fastapi = {extras = ["standard"], version = "^0.112.0"}
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.31"
pytest = "^8.3.2"
punq = "^0.7.0"
black = "^24.8.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
